Booking management

+++ src/main/kotlin/nl/avans/routes/cars/Cars.kt
@@ -2,6 +2,7 @@ package nl.avans.routes.cars
 
 import io.ktor.http.*
 import io.ktor.server.application.*
+import io.ktor.server.plugins.*
 import io.ktor.server.request.*
 import io.ktor.server.response.*
 import io.ktor.server.routing.*
@@ -13,32 +14,52 @@ fun Application.configureCarRouting(carDAO: CarDAO) {
     routing {
         // Create car
         post("/car") {
-            val car = call.receive<Car>()
-            val id = carDAO.create(car)
-            call.respond(HttpStatusCode.Created, id)
+            try {
+                val car = call.receive<Car>()
+                val id = carDAO.create(car)
+                call.respond(HttpStatusCode.Created, id)
+            } catch (e: BadRequestException) {
+                call.respond(HttpStatusCode.BadRequest, "Car values are incorrect!")
+            }
         }
         // Read car
         get("/car/{id}") {
-            val id = call.parameters["id"]?.toInt() ?: throw IllegalArgumentException("Invalid ID")
-            try {
-                val car = carDAO.read(id)
-                call.respond(HttpStatusCode.OK, car)
-            } catch (e: Exception) {
-                call.respond(HttpStatusCode.NotFound)
+            val id = call.parameters["id"]?.toIntOrNull()
+            if (id != null) {
+                try {
+                    val car = carDAO.read(id)
+                    call.respond(HttpStatusCode.OK, car)
+                } catch (e: Exception) {
+                    call.respond(HttpStatusCode.NotFound)
+                }
+            } else {
+                call.respond(HttpStatusCode.BadRequest, "Invalid ID")
             }
         }
         // Update car
         put("/car/{id}") {
-            val id = call.parameters["id"]?.toInt() ?: throw IllegalArgumentException("Invalid ID")
-            val car = call.receive<Car>()
-            carDAO.update(id, car)
-            call.respond(HttpStatusCode.OK)
+            val id = call.parameters["id"]?.toIntOrNull()
+            if(id != null) {
+                try {
+                    val car = call.receive<Car>()
+                    carDAO.update(id, car)
+                    call.respond(HttpStatusCode.OK)
+                } catch (e: BadRequestException) {
+                    call.respond(HttpStatusCode.BadRequest, "Bad request, check the values")
+                }
+            } else {
+                call.respond(HttpStatusCode.BadRequest, "Invalid ID")
+            }
         }
         // Delete car
         delete("/car/{id}") {
-            val id = call.parameters["id"]?.toInt() ?: throw IllegalArgumentException("Invalid ID")
-            carDAO.delete(id)
-            call.respond(HttpStatusCode.OK)
+            val id = call.parameters["id"]?.toIntOrNull()
+            if (id != null) {
+                carDAO.delete(id)
+                call.respond(HttpStatusCode.OK)
+            } else {
+                call.respond(HttpStatusCode.BadRequest, "Invalid ID")
+            }
         }
     }
 }
